Auto-format markdown on enter press on the main thread
Blocking the main thread is bad, but a delay in showing the formatted
note will look worse. I _believe_ parsing of text shouldn't be expensive.
Parsing is skipped anyway if a fenced code block or a list item wasn't
detected. The only part I'm concerned of is the usage of regex for long
notes. Can optimize it in the future if it proves to be a bottleneck.